backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
sqlalchemy[asyncio]>=2.0.36
asyncpg>=0.30.0
alembic>=1.14.0
pydantic-settings>=2.6.0
python-jose[cryptography]>=3.3.0
anthropic>=0.40.0
openai>=1.58.0
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-community>=0.3.0
pgvector>=0.3.0
google-api-python-client>=2.100.0
httpx>=0.27.0
sse-starlette>=2.0.0
python-multipart>=0.0.9
redis>=5.0.0
psycopg2-binary>=2.9.0
